Add files via upload
Due in part to working this weekend, I was only able to get as far as having it return the letters in the word.  I understand that the guesses need to be stored in a list, but unsure how to do that, as .append seems to only add at the end of the list string.